% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estat.R
\name{read_estat_file}
\alias{read_estat_file}
\title{Read files downloaded from e-stat}
\usage{
read_estat_file(file, statsCode = "00400004", id = "")
}
\arguments{
\item{file}{path}

\item{statsCode}{statsCode}

\item{id}{Not currently used.}
}
\description{
Read files downloaded from e-stat
}
