% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_map.R
\name{read_estat_map}
\alias{read_estat_map}
\title{Read e-Stat aggregation unit boundary data}
\usage{
read_estat_map(file, type = "aggregate_unit", remove_cols = TRUE)
}
\arguments{
\item{file}{Path to downloaded e-Stat shape file}

\item{type}{Currently, only "aggregate_unit" is used.}

\item{remove_cols}{Whether or not to remove redundant columns.
When \emph{TRUE} (the default), the following columns are removed (See details).
These columns can be substituted or sought with values from other columns.
\itemize{
\item S_AREA
\item KAxx_, KAxx_id
\item KEN, KEN_NAME
\item DUMMY1
\item X_CODE, Y_CODE
}}
}
\description{
The GIS data downloaded from e-Stat is read and converted into an easy to process \link[sf:sf]{sf::sf} format.
You can use \code{\link[=download_stat_map]{download_stat_map()}} to download the data.
}
