% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estat_meshpop.R
\name{read_estat_meshpop}
\alias{read_estat_meshpop}
\title{Read e-Stat mesh population census data}
\usage{
read_estat_meshpop(file, mesh_size = 0.5)
}
\arguments{
\item{file}{Path to downloaded e-Stat text file}

\item{mesh_size}{Currently, only 0.5 (500m mesh)}
}
\description{
Read e-Stat mesh population census data
}
