% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_map.R
\name{download_stat_map}
\alias{download_stat_map}
\title{Download Japan prefecture map from e-Stat}
\usage{
download_stat_map(
  prefcode,
  exdir = ".",
  dest = TRUE,
  .survey_id = c("A002005212015")
)
}
\arguments{
\item{prefcode}{The JIS-code for prefecture and city identical number.
If prefecture, must be from 1 to 47.}

\item{exdir}{The directory to extract files to (the equivalent of
    \code{unzip -d}).  It will be created if necessary.}

\item{dest}{If \emph{TRUE}, to unzip downloaded files.}

\item{.survey_id}{survey id (A00200521YYYY is small area,
D00200521YYYY is did area)}
}
\description{
Download Japan prefecture map from e-Stat
}
\seealso{
\url{https://www.e-stat.go.jp/gis}
}
