% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_map.R
\name{download_stat_map}
\alias{download_stat_map}
\title{Download Japan prefecture map from e-stat}
\usage{
download_stat_map(prefcode, exdir = ".", dest = TRUE)
}
\arguments{
\item{prefcode}{jis code for prefecture and city identifical number. If prefecture, must be from 1 to 47.}

\item{exdir}{The directory to extract files to (the equivalent of
    \code{unzip -d}).  It will be created if necessary.}

\item{dest}{If TRUE, to unzip downloaded files.}
}
\description{
Download Japan prefecture map from e-stat
}
