% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estat_meshpop.R
\name{read_estat_meshpop}
\alias{read_estat_meshpop}
\title{Read e-Stat mesh population census data}
\usage{
read_estat_meshpop(file, is_long = FALSE, mesh_size = 0.5)
}
\arguments{
\item{file}{Path to downloaded e-Stat text file,
such as \verb{tblT000847H\{meshcode\}.txt}}

\item{is_long}{If \emph{TRUE}, make the population category a single variable
and make the values of the counterpart population independent.}

\item{mesh_size}{Currently, only 0.5 (500m mesh)}
}
\description{
Read e-Stat mesh population census data
}
